refactor(agentPresenter): move streaming module
