+++ src/main/presenter/agentPresenter/loop/toolCallHandler.ts
@@ -9,7 +9,7 @@ import type {
 } from '@shared/presenter'
 import { nanoid } from 'nanoid'
 import type { MessageManager } from '../../sessionPresenter/managers/messageManager'
-import type { GeneratingMessageState } from '../../sessionPresenter/streaming/types'
+import type { GeneratingMessageState } from '../streaming/types'
 import type { CommandPermissionService } from '../../permission/commandPermissionService'
 
 interface PermissionRequestPayload {
