+++ src/main/presenter/agentPresenter/streaming/contentBufferHandler.ts
@@ -1,7 +1,7 @@
 import { eventBus, SendTarget } from '@/eventbus'
 import { STREAM_EVENTS } from '@/events'
 import { finalizeAssistantMessageBlocks } from '@shared/chat/messageBlocks'
-import type { MessageManager } from '../managers/messageManager'
+import type { MessageManager } from '../../sessionPresenter/managers/messageManager'
 import type { GeneratingMessageState } from './types'
 
 export class ContentBufferHandler {
