+++ src/main/presenter/agentPresenter/streaming/types.ts
@@ -1,5 +1,5 @@
 import type { AssistantMessage } from '@shared/chat'
-import type { PendingToolCall } from '../../agentPresenter/message/messageBuilder'
+import type { PendingToolCall } from '../message/messageBuilder'
 
 export interface GeneratingMessageState {
   message: AssistantMessage
