+++ src/main/presenter/searchPresenter/handlers/searchHandler.ts
@@ -3,7 +3,7 @@ import { finalizeAssistantMessageBlocks } from '@shared/chat/messageBlocks'
 import type { MODEL_META, SearchResult } from '@shared/presenter'
 import { nanoid } from 'nanoid'
 import { buildUserMessageContext } from '../../agentPresenter/message/messageFormatter'
-import type { GeneratingMessageState } from '../../sessionPresenter/streaming/types'
+import type { GeneratingMessageState } from '../../agentPresenter/streaming/types'
 import { BaseHandler, type ThreadHandlerContext } from './baseHandler'
 
 interface SearchHandlerOptions {
