+++ src/main/presenter/sessionPresenter/index.ts
@@ -22,12 +22,12 @@ import { eventBus } from '@/eventbus'
 import { TAB_EVENTS, CONVERSATION_EVENTS } from '@/events'
 import type { ISessionPresenter } from './interface'
 import { MessageManager } from './managers/messageManager'
-import type { GeneratingMessageState } from './streaming/types'
-import { ContentBufferHandler } from './streaming/contentBufferHandler'
+import type { GeneratingMessageState } from '../agentPresenter/streaming/types'
+import { ContentBufferHandler } from '../agentPresenter/streaming/contentBufferHandler'
 import { ToolCallHandler } from '../agentPresenter/loop/toolCallHandler'
-import { LLMEventHandler } from './streaming/llmEventHandler'
+import { LLMEventHandler } from '../agentPresenter/streaming/llmEventHandler'
 import { SearchHandler } from '../searchPresenter/handlers/searchHandler'
-import { StreamGenerationHandler } from './streaming/streamGenerationHandler'
+import { StreamGenerationHandler } from '../agentPresenter/streaming/streamGenerationHandler'
 import { PermissionHandler } from './permission/permissionHandler'
 import { UtilityHandler } from './utility/utilityHandler'
 import { CommandPermissionService } from '../permission/commandPermissionService'
