+++ src/main/presenter/sessionPresenter/permission/permissionHandler.ts
@@ -14,9 +14,9 @@ import {
   buildPostToolExecutionContext,
   type PendingToolCall
 } from '../../agentPresenter/message/messageBuilder'
-import type { GeneratingMessageState } from '../streaming/types'
-import type { StreamGenerationHandler } from '../streaming/streamGenerationHandler'
-import type { LLMEventHandler } from '../streaming/llmEventHandler'
+import type { GeneratingMessageState } from '../../agentPresenter/streaming/types'
+import type { StreamGenerationHandler } from '../../agentPresenter/streaming/streamGenerationHandler'
+import type { LLMEventHandler } from '../../agentPresenter/streaming/llmEventHandler'
 import { BaseHandler, type ThreadHandlerContext } from '../../searchPresenter/handlers/baseHandler'
 import { CommandPermissionService } from '../../permission/commandPermissionService'
